+++ sdk/core/azure-core-xml/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 1.0.0-beta.1 (Unreleased)
 
-* Initial release.
+- Initial release.